lab14 - fixes

+++ lab14/Flat.java
@@ -4,10 +4,10 @@ import java.time.LocalDate;
 
 public final class Flat extends Property {
 
-    private double flatNumber;
-    private double floorNumber;
+    private int flatNumber;
+    private int floorNumber;
 
-    public Flat(String street, int houseNumber, String city, String postalCode, double area, double price,
+    public Flat(String street, int houseNumber, String city, String postalCode, double area, int price,
             LocalDate dueDate, int flatNumber, int floorNumber) {
         super(street, houseNumber, city, postalCode, area, price,
                 dueDate);
@@ -24,19 +24,19 @@ public final class Flat extends Property {
         return super.toString() + ", numer mieszkania: " + this.flatNumber + ", numer piętra: " + this.floorNumber;
     }
 
-    public double getFlatNumber() {
+    public int getFlatNumber() {
         return flatNumber;
     }
 
-    public void setFlatNumber(double flatNumber) {
+    public void setFlatNumber(int flatNumber) {
         this.flatNumber = flatNumber;
     }
 
-    public double getFloorNumber() {
+    public int getFloorNumber() {
         return floorNumber;
     }
 
-    public void setFloorNumber(double floorNumber) {
+    public void setFloorNumber(int floorNumber) {
         this.floorNumber = floorNumber;
     }
 }
